Add stream support
Now pack can receive any number of arguments returning a stream
of the serialized data (on 0 objects, returns the empty string).
unpack deserialized any stream of objects, returning the result
of the whole stream deserialization (this, the number of return
values of unpack is equal to the number of objects unpacked).
If given the empty string, returns nothing.

I also removed all warnings and made the code C++ compatible.

[untested at this point]