Add better type checking when Lua 5.3 is available
Lua 5.3 adds an actual Integer type, which by default
will be a 64-bit long long.  Currently, Lua {5.1,5.2}
only has a Number type which by default is a 64-bit
double.

Note: Lua can be custom compiled to re-define Integer
to be 32 bits (as well as Number to be just a float), but
that should be rare and the interal cmsgpack conversion
routes store integers in the most compact representation
discernable.

This also uses the existing lua_tointeger function instead of
casting a lua_Number to an int64_t directly.