Unpack integers as integers instead of numbers
This fixes Lua 5.3 compatability and works fine on Lua 5.1 too.

Lua 5.3 has an actual 64-bit Integer type, so we need to use
_pushinteger instead of _pushnumber so our unpacked integers
don't convert to floating point values.

For older versions, lua_pushinteger casts input to the default numeric type
and behavior is unchanged.

Lua 5.3 also has an unsigned integer type, so I added a compatibility
macro to allow us to properly use lua_pushunsigned on 5.3+
with a fallback to the regular lua_pushinteger on other versions.