5.3 compat: short functions to long functions
checkint -> checkinteger
optint -> optinteger

Also removes `< 503` check for creating pushunsigned.

Older lua versions had redundant functions, but one with a shorter
name than the other.  Newer Lua versions only retain the long version.

one is an alias of the other, but one was removed in 5.3+, so it's
better to use the common function across all versions.